Youtube_To_Blog_Using_CrewAI/requirements.txt
crewai
crewai_tools
load-dotenv
langchain-huggingface